Add Vertex Handles and Bezier Interpolation
add vertex handles and ability to use bezier interpolation in
conjunction with those handles to draw curves. handles can also be
drawn, which is useful for debugging or spline manipulation. handles can
be constrained in a couple of useful ways (mirrored or copied angle).
primitive type can now be set to any sfml primitive type.