Ring - add bounds, sector, area
add local and global bounds methods.
add ability to output the area of the ring.
add ability to specify a sector of the circle/ring. the values are
ratios of the full ring (range 0 to 1). offset is the start point from
angle 0, size is the end point from the offset.