Progress Bar - update to v1.1
update to v1.1.
allow textures to be applied to the bar and/or the background.
use vertex array for bar instead of rectangle shape to allow float
values for the texture co-ordinates.
rename "enabled" setters and getters to "show":
setBarEnabled to setShowBar.
add example and image resource for repeating gradient background
texture.