Console Screen - update to v1.2
add paintAt methods, which match printAt methods, that prints colour(s)
without affecting the values of the cells. The string/char parameter of
printAt is replaced by an unsigned int in paintAt that represents the
length of the painting (in cells)
add buffers/clipboards. copy and paste methods can capture and recall
the entire screen's contents, or just a rectangular section of it, which
can be offset when pasted. multiple buffers can be
stored/manipulated/removed
add cell information retrieval methods (getters) that supply the value,
colour, background colour, or entire cell at the requested location
remove use of "greyscale" and use only "grayscale"