Tile Map - update to v1.1
added "smooth scroll" to offset the render's position to more accurately
reflect the camera's position (round to pixel of the view rather than
pixel of the render) - similar effect to video stabilization.
added ability to specify a texture offset for when the tileset is a
sub-texture.
added ability to specify a "camera tile target", which allows the camera
position to point to somewhere other than the top-left of the render.
added ability to get the level position of the tile that is at the
co-ordinate specified. takes into account any transformations.
added ability to get the value of the tile that is at the co-ordinate
specified. takes into account any transformations.
added some missing consts to parameters that should be constant.
updated update() functions to use the correct data type for level value.