Console Screen - update to v1.5
Update version one of Console Screen to its latest state in preparation
of additional inclusion of Console Screen version 2.
Note that its filenames have been appended with "Old" and is included as
such in the SelbaWard.hpp inclusion list.
Note also that the class name is now ConsoleScreenV1 to avoid any
clashing with the latest version (2), which uses "ConsoleScreen"