add Console Screen v2
initial commit of version 2. version 1 has been renamed in advance so
that they can both be included in Selba Ward concurrently.
includes also a file that contains a number of enums that can be used to
specify a palette index for each of the palettes in Console Screen.
Selba Ward's global inclusion list includes the palette enum file but
Console Screen doesn't include it automatically so it must be included
manually if you manually include specific Selba Ward objects.