Starfield - add setColor
add setColor to allow changing the colour of the stars without having to
re-generate them. The depth is determined by the alpha so the stars keep
their alpha; the alpha of the passed colour is ignored.
removed all regenerate methods that include colour parameters as they
can now be done without the colour and then the colour can be set
separately, if required. this reduces the number of possible parameter
combinations in the regenerate method.
also small fix: randomAlpha is now cast before storing into the vertex
colour's alpha component.