Console Screen - update to v2.1
updated to v2.1
added method to retrieve "perfect size", which is the size where each
pixel of the texture would be displayed as one pixel.
scrolling can now be confined to a rectangular region (selection).
colours can now be "cycled", which is shifting (or scrolling) the
colours in the palette. Cycling always wraps around.
Cursor commands (Newline, Backspace and Delete) act differently if the
current printing properties have stretch set to "Both". In this case,
Newline moves down two rows instead of just one whereas Backspace and
Delete both now clear the cell below the usual one that is cleared - if
possible.
fixed printing in bottom-right cell not initiating automatic scrolling.
fixed setting a palette colour not updating the screen.