Console Screen - update to v2.2
added ability to stream single characters directly rather than first
having to convert to a string.
buffers can now be inserted and resized as well as have their cells
manipulated directly without the need to copy from the current screen.
fixed printing error with double-height printing that caused the lower
line to not be printed when it should (stretch type of "Both")
added the specification of storage type of forward-declared enums.