added control macro for deprecated SFML features
added macro to control use of deprecated SFML features.
USE_SFML_PRE_2_4 is defined when the current SFML is version 2 but
before version 2.4. If this is defined, Selba Ward uses the versions
that were deprecated in v2.4 but uses the newer versions if the current
SFML version is 2.4+.
updated all classes that use deprecated SFML or newer SFML to use the
macro to control which to use.