Spline - update to v1.2
added ability to automatically close spline. in this case, the first
vertex is treated as though it also exists after the final vertex, which
includes the interpolation from final vertex to the first vertex. thick
spline's angle at the first/last vertex takes into account both sides
when closed.
vertices' positions (sf::Vector2f) can now be provided directly through
the constructor using the initializer list. the number of vertices in
the spline is determined by the number of positions provided in the
list.
added ability to retrieve the interpolated positions of the spline (an
assert is included to check the validity of the requested position) as
well as the total number of interpolated positions. note that this total
counts the final vertex when the spline is open (i.e. not closed) but
does not count it when it is closed. however, the interpolated positions
are still counted. therefore, cycling from 0 to
(totalInterpolatedPositions - 1) would cycle through all interpolated
positions on the closed spline.
setting the colour didn't affect the thick spline - only the thin
spline; this has been fixed.
during the update method, the vertices' vectors are resized once to
reserve the vertices needed. the calculation was incorrect and has now
been corrected.