NinePatch - update to v1.4
added ability to use rectangle to specify a part of a texture to use
instead of requiring the entire texture.
renamed getContentArea to getGlobalContentArea as this method retrieves
the content area after transformations.
added getLocalContentArea to get the content area before
transformations.
added ability to test if a point is inside the transformed content area.
manually resetting the size now matches the texture rectangle instead of
the entire texture.
added ability to reset texture rectangle to entire texture when setting
texture (note that this defaults to true).