Spline - update to v1.3
added per vertex thickness and colour. the global thickness and colour
are still present and are multiplied by the vertex values, which are
linearly interpolated throughout the spline. since they are multiplied,
the default values for a vertex's thickness is 1 and for a vertex's
colour is white (255, 255, 255, 255).
note that using different thicknesses may cause slight artifacts around
very sharp corners.
added ability to get the tangents of the interpolated positions. that
is, a unit vector representing the current 'direction' of the line at
that point. tangents are calculated during an update always (whether the
spline is thick or thin) so can be retrieved trivially.
added ability to retrieve the normal at interpolations position. this
normal is simply a unit vector representing the tangent rotated 90
degrees anti-clockwise.
added ability to retrieve the interpolated spline thickness at
interpolated positions. this thickness is the interpolation of the two
surrounding vertices' thicknesses. does not pre-multiply by global
thickness.
added ability to retrieve the "thickness correction scale" at an
interpolated position. this correction is the amount that the thickness
is scaled so that the thickness is constant regardless of curvature.
this scale is highest at sharp corners.
refactored the section that calculates the thick vertices as it now uses
the precalculated tangents.