Bitmap Text - update to v1.1.1
a null font is now considered a valid state and checks to see if it's
null. setBitmapFont is therefore now able to take no parameters to
nullify the font.
vertices are now updated when a font is set (including changed).
setting the color no longer requires an entire update for the vertices.
added "priv_" prefix to private methods as per convention.
reformatted the constructor's member initialisation as per convention.
if a font is null, updating the vertices simply clears the vertices and
resets the bounds to zero.
if no vertices exist (a null font is one case that can cause this), the
draw method is immediately exited with no code being executed.