Elastic Sprite - update to v1.1
added ability to use perspective interpolation.
either interpolation type (bi-linear - the original - or perspective)can
be activated at any time; this disabled the opposite interpolation.
perspective interpolation calculates weights per vertex that must be
interpolated as with any other vertex attribute. unfortunately, SFML has
a strict and rigid vertex format so an extra float could not be passed.
this led to using the RGB components of the color attribute in the
vertex attribute instead to store the weight value (the alpha is not
altered). due to this, the RGB part of the color is ignored when using
perspective interpolation (the alpha is still used).
perspective interpolation is then performed using bespoke vertex and
fragment shaders that make use of the vertex weight (stored in the
colors' RGB).

removed "T" from template line in forward declaration of SFML classes
(i.e. template <class T> became template <class> since the name here is
not used.