Console Screen - update to v2.4
added ability to get a cell location at a given co-ordinate as well as
the (transformed) co-ordinate of a cell location. cell location here is
not integers so this allows any location within the cell.

added new "Number" type to allow numbers to be streamed and
automatically printed (std::to_string)

added new 8-colour RGB palette - all variations of the three components
(RGB) being fully off or fully on.

added ability to get the attributes bitmask from a cell. this is an
unsigned int that represents all attributes values (as Affect values so
this value will always be between 129 and 255 except for zero -
attributes in Affect are in the first "nibble" of the least significant
byte).

added ability to set all texture values in a single function. this
allows setting up the texture to be done 'in one go'.

added ability to change the clear value - the value to which a cell's
value is changed when it is 'cleared' - from zero. if a char is used, it
is also mapped using character mapping.

added ability to change the multiplier that is used for colours when the
dark attribute is set.

bright attribute, which defaulted to true/on, is now dark attribute,
which defaults to false/off. this means that the default attributes are
now all false, simplifying default attributes. moved the
"makeColorUnBright" function from the anonymous namespace to the class'
private methods, now named "priv_makeColorDark", as it uses the class'
dark multiplier member.

initial value for number of texture tiles per row is now 16 instead of
8. the default value for the number of texture tiles per row parameter
is now 16 instead of 1.

the function that allows cycling palette using a range has had its
parameters re-arranged, putting the 'amount to cycle' at the end,
allowing it to be left off (defaults to one) and just the range
specified.

char values above 127 in a string are no longer ignored; the negative
values of char are now converted to their positive values from 128 and
above, allowing all 256 values of char (which is signed) to be used and
passed as a cell value (which is unsigned).

changed default foreground and cursor colours. those default colours
(although the actual colours could be changed) were hard-coded to '15'
and causes problems when using a palette with fewer than 16 colours,
especially when using automatic updates. these default colours are now
set to the colour command Contrast, which contrasts the background at
best possible. of course, the colours can still be changed as normal.

changing stretch type no longer (when automatic update is enabled)
updates the entire screen, which could unnecessary delays and unexpected
when stretch is changed a lot of times every loop. automatic updates
only update the cells that are modified and stretch didn't previously
follow this rule.

removed unnecessary use of abs.

code-style modifications:
moved 'tiny' methods (public only) to header as inline methods.
added "int" keyword to "long int" as per convention.
values for Affect enum are now written in hexadecimal, making it clearer
that all Affect values are within a byte (0 - 255) and the first (most
significant) nibble is for attributes and the second (least significant)
nibble is for the rest (value, colours and stretch).