Elastic Sprite - update to v1.2 ( fixes #22 )
The main update here is the fix of the perspective interpolation errors that occur using large offsets for the texture rectangle. It is likely to also apply to just using very large textures.

Rewrote the perspective shaders to use a different method to pass the weights to the vertices within the shaders. This allows them to be passed as floats and that removes the "approximation errors" made by passing as separate integers.
The new vertex shader requires multiple uniforms to be set for it. The shader then uses encoded colour values to determine which vertex it current is and sets the correct weight for that vertex to the texture co-ordinate's z position.
The colours are stored before drawing and coding, then restored immediately after drawing. This allows the vertex colours to be used; this wasn't possible in the previous version. Note, though, that the interpolation of the vertex colours is the 'normal' way; as two separate triangles.

Some slight code tweaks too:
updated year.
removed the test to see if vertexIndex is zero or above; it's unsigned so this was redundant.
after updating the vertices in the draw function, the states.transform is set to the identity transform since all vertices have already been transformed; there is no need to transform them again in target.draw.
target.draw always uses 4 vertices and is now hard coded. it also is included in each branch of the ifs since one of them requires extra work after it using local data.
removed all the integer colour coding.
textureSize is set as a float vector since its values are always used as floats anyway.