Spline - update to v1.5
update Spline to version 1.5

add 2 new corner types, both are variations on the point type: PointLimit and PointClip. both use a new value called "MaxCornerPointLength" and use a modified corner if the point corner is longer than this value. Limit resizes the point to this maximum value whereas clip converts the corner to a bevel-style if the point max is surpassed. if the maximum length hasn't been reached, it forms a standard point corner.
note that limiting the point to less than the thickness pulls in the corners. limiting it to 0 shows the spline as connected rectangles (with no corner joins).

removed some commented-out code that is no longer needed.

reorganised some comments to be more clear and precise.

reorganised some of the "update output vertices" code to allow some values to be used more than they were previously. this includes keeping the original tangent unit before transformation for the point corner.

combined PointClip, PointLimit and Bevel code in "update output vertices" as they use a similar approach.

fixed the end cap. the vertices within each vertex pair were added in the wrong order, which formed an overlapping triangle. this overlap is now gone.