Spline - update to v1.6.0
added ability to rotate entire spline around any position as its origin.

added ability to add vertices of another spline to the front or back (before or after the current spline).

added ability to "connect" a spline to another spline. this allows the front or back vertex of a spline to be moved to match the front or back of another spline. when moving this vertex, the entire spline can be moved with it. also, the spline can be rotated to match the orientation of the vertex of the other spline to which it will be connected. rotation and moving are on by default and can be disabled by the boolean parameters. they can also be used together or separately of each other.

added ability to add another spline and connect it in one go. takes the same booleans as the connect methods.

note that when adding the connection is from the other spline "to" the current one.
when only connecting, it is from the current one "to" the other one.

added a flag to enable/disable the random normal offsets. this removes unnecessary calculations using random numbers when not in use). note that this is disabled by default for standard use so any use of the random normal offsets will now need to enable this specifically.

added a reserve call when adding multiple vertices.

re-arranged some of the code in the vertices update method to allow the random offset normal flag to control the correct parts in single blocks.