Elastic Sprite - update to v1.3.0
added new ability to assign a flip 'flag' to allow to flip the texture in both/either horizontal and/or vertical direction.

works with the texture with either shader and also without the shader.

now stores two texture rectangles: the base one (the one chosen by the user) and the actual one (the one used to get the co-ordinates from the texture). the actual one is based on the base one and is updated whenever the vertices are updated.