add Starfield 3D
add new drawable: Starfield 3D
similar to Starfield, it creates a randomised set of stars that can be moved with parallax.

however, this version uses actual 3D projection and allows Z movement as well.

you can specify the front and back colours as well as the front and back scale (of the star shape) and they are automatically "tweened" between them.

included is the ability to create a "star template" that is the shape of each star. it can be any shape or colour. colours are multiplied with the colour (from the front and back colours).

you can "move" in all three dimensions and pan in two dimension. "roll" rotation is not included; you can simply use the sf::Drawable's rotation feature for that.