Polygon-update to v1.2
in addition to BasicEarClip, which triangulates simple polygons, EarClip is now provided, which can triangulate simple polygons that contain any number of simple polygon holes.

some restrictions:
polygons vertices must be provided in an anti-clockwise order.
holes' vertices must be provided in a clockwise order.
all vertices (polygon and holes) must be provided together.
holes' vertices must be provided after all of the polygon vertices.
an index (per hole) must be provided to mark the start of that hole.