Polygon-update to v1.2.3
now throws a Selba Ward exception showing an error code instead of attempting to access an empty vector (during the update method).

this allows both:
reporting an error for investigation and
catching the error suring runtime to avoid invalid use.

to catch this exception, simply put the update() method in a try block. remember to reset any changes before attempting to update again (e.g. if you've changed a vertex position before the update that failed, change the vertex position back to what it was before)