add SpriteBatch v1.0.0
add Sprite Batch, v1.

stores multiple sprites internally and creates a single vertex array from them when drawing. can avoid global updates in basic situations.

the stored sprites can be manipulated indirectly only via class methods. this allows only updating what is required.

if an order function exists (to sort pointer to sprites), it is used to order the sprites when updating; this forces global update to always be active. if no function exists, it will order them according to their storage and global updates are not required.

if a manual order (a vector of sprite indices) exists, this override the order function if it exists (but does not clear it). the manual order is used for the first part of the vector and any missing indices will follow in original order. if the manual order is cleared, the order function will be re-activated if it exists, otherwise it'll use the normal method.

setting the order function clears the manual order but clearing the order function does not clear the manual order.

neither setting nor clearing the order clears the order function (but if a manual order exists, it overrides the order function).

clearAllOrdering() is provided to clear both the order function and the manual order at once.

all standard SFML sprite methods are present but require a sprite index to be included (to identify which sprite to apply it to).

if "move", "rotate" and "scale" are used without a sprite index, it will be applied to all sprites.