update Polygon to v1.3
update Polygon to v1.3, adding some extra functionality.

adds ability to manually set triangle limit: maximum amount of triangles that can be created before the update is halted. the default is 10000 (this limit was previously hard-coded as 100).

adds ability to add a "wireframe" mesh over the top of the polygon. this is drawn using separate lines and is rebuilt (updated) whenever the polygon is updated. it's based on the final triangles after triangulation so will continue to work when other methods are added.

the wireframe can be set to be shown or not and its colour can also be set. if it's set to not be shown, it will clear it (as well as not build it or draw it) during the update. note that this means that update will need to be called if this is modified.

in addition, casts have been corrected: now casts to the long double it should be. also added const where a parameter is not expected to change, and removed some unused (commented-out) code.