update SpriteBatch to v1.0.1
update Sprite Batch to version 1.0.1, fixing some issues moving to SFML 3.

since SFML sprites can no longer be constructed without a texture, we need to create a "dummy" empty texture that use to construct all of the sprites in the vector as well as the default one passed as a parameter.

also feed the angles to SFML angles as degrees as was the original default. note that this is likely to be updated in the future so that it's closer to SFML's interface for its sprites.