update Sprite3d to v1.3
updates Sprite 3D to v1.0.0

adds ability to "setTexture" with no parameters, allowing no texture to be set. this allows textures to be disabled and the solid colour used instead. note that this can be done per face so one side can be textured and the other can be a solid colour, for example.

note also, that this means that disabling the texture does not disable the subdivisions and 3D calculations even though they are likely unnoticeable. changing the subdivisions and mesh size when disabling/enabling textures can increase efficiency and optimise performance, if required.