Update Line to v1.3
updates Line to v1.3, which removes PointIndex type as it was only used as a return value for getStartIndex and getEndIndex. These now just simply return 0 and 1 respectively. PointIndex was not even used to specify the point index; it took a value of 0 or 1 so it was useless.
Also updated code style.